backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
python-multipart==0.0.6
PyJWT==2.8.0
cryptography==42.0.0

# Git providers
PyGithub==2.1.1
python-gitlab==4.4.0

# LLM backends
google-cloud-aiplatform==1.38.1
boto3==1.34.21
openai==1.10.0

# Database
supabase==2.3.0

# Utils
aiofiles==23.2.1
pyyaml==6.0.1
